requirements.txt
chromadb
sentence-transformers
openai
numpy
python-dotenv
